requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
chromadb>=0.4.0
markdown>=3.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
pypdf>=3.17.0
python-dotenv>=1.0.0

